% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{The genotype matrix of the target population}
}
\value{

}
\description{
This function is used to normalize genotypes matrix
}
