% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLMulti.R
\name{pheno_generation}
\alias{pheno_generation}
\title{pheno_generation}
\usage{
pheno_generation(Ne, Na, Nt, Za, Ze, Zt, ratio, rho, h2)
}
\arguments{
\item{Ne}{The number of samples for the informative auxiliary population used
for simulating summary statistics.}

\item{Na}{The number of samples for the training target population used for
simulating summary statistics.}

\item{Nt}{The number of samples for the individual-level data for the target
population}

\item{Za}{Genotypes of the training samples.}

\item{Ze}{Genotypes of the informative auxiliary samples.}

\item{Zt}{Genotypes of the testing samples.}

\item{ratio}{The proportion of causal SNPs to common SNPs.}

\item{rho}{The genetic correlation between these two populations.}

\item{h2}{The value of heritability.}
}
\value{
A list of phenotypes according to the input genotypes.
}
\description{
This function is used to generate phenotypes of simulation
  data.
}
\details{
This function will return the phenotypes of two samples, in which
  most causal variants are shared. There is a difference between the target
  samples and the informative auxiliary samples in the genetic architecture,
  which causes estimation bias.
}
