% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_tf.R
\name{PRS_tf}
\alias{PRS_tf}
\title{PRS_tf}
\usage{
PRS_tf(
  ss_e,
  ss_a,
  info_LD,
  tar_LD,
  X,
  PATH_TO_EUR,
  PATH_TO_REF,
  PATH_TO_TEST,
  tar_pheno,
  sample_r = NULL,
  sample_t = NULL,
  cluster = NULL
)
}
\arguments{
\item{ss_e}{GWAS of the informative auxiliary samples.}

\item{ss_a}{GWAS of the target samples.}

\item{info_LD}{LD matrix in Lassosum for the informative auxiliary samples}

\item{tar_LD}{LD matrix in Lassosum for the target samples}

\item{X}{The individual-level data of the target samples (unnormalized).}

\item{PATH_TO_EUR}{The data path of reference panel for the informative auxiliary population}

\item{PATH_TO_REF}{The data path of reference panel for the target population.}

\item{PATH_TO_TEST}{The data path of reference test panel for the target population.}

\item{tar_pheno}{The phenotype of the target population.}

\item{sample_r}{Participants to keep from the reference panel}

\item{sample_t}{Participants to keep from the test panel.}

\item{cluster}{A \code{cluster} object from the \code{parallel} package for parallel computing.}
}
\value{
A list of multiethnic PRS calculated by TL-Multi.
}
\description{
This function is used to run TL-Multi which is applied to conduct multiethnic PRS prediction
}
\details{
This function will return the multiethnic PRS calculated by TL-Multi
}
