% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS_tf.R
\name{PRS_tf}
\alias{PRS_tf}
\title{PRS_tf}
\usage{
PRS_tf(
  ss_tl,
  pheno,
  ref.bfile,
  test.bfile,
  LDblocks,
  keep.test = NULL,
  keep.ref = NULL
)
}
\arguments{
\item{ss_tl}{The GWAS data for TL-Multi}

\item{pheno}{The phenotype of the target population}

\item{ref.bfile}{The reference panel for the target population}

\item{test.bfile}{The testing panel for the target population}

\item{LDblocks}{The LD block for the target population}

\item{keep.test}{Participants to keep from the testing dataset}

\item{keep.ref}{Participants to keep from the reference panel}
}
\description{
This function is used to apply TL-Multi to conduct multiethnic PRS prediction
}
