% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tl.R
\name{ss_tl}
\alias{ss_tl}
\title{PRS_tf}
\usage{
ss_tl(v_e, out_e, ss_a, XTX = XTX)
}
\arguments{
\item{v_e}{The validation result of Lassosum for the auxiliary informative population}

\item{out_e}{The Lassosum result of the auxiliary informative population}

\item{ss_a}{The GWAS data for the target population}

\item{XTX}{The LD region of the target population}
}
\value{
A list of multiethnic PRS calculated by TL-Multi.



snp = snp,
}
\description{
This function is used to generate the GWAS data for TL-Multi which were applied to conduct multiethnic PRS prediction
}
\details{
This function will return the multiethnic PRS calculated by TL-Multi
}
